API-3: refactoring to use user relationships instead of onfere the user_id on the creation
