+++ app/Http/Controllers/Question/StoreController.php
@@ -3,20 +3,19 @@
 namespace App\Http\Controllers\Question;
 
 use App\Http\Resources\QuestionResource;
-use App\Models\Question;
 use App\Http\Controllers\Controller;
-use App\Http\Requests\Quesstion\StoreRequest;
-use Symfony\Component\HttpFoundation\Response;
+use App\Http\Requests\Question\StoreRequest;
 
 class StoreController extends Controller
 {
     public function __invoke(StoreRequest $request)
     {
-        $question = Question::create([
-            'question' => $request->question,
-            'status'   => 'draft',
-            'user_id'  => auth()->user()->id,
-        ]);
+        $question = user()
+            ->questions()
+            ->create([
+                'question' => $request->question,
+                'status'   => 'draft',
+            ]);
 
         return QuestionResource::make($question);
     }
