+++ app/Support/functions.php
@@ -1,0 +1,9 @@
+<?php
+
+if (!function_exists('user')) {
+
+    function user(): ?\App\Models\User
+    {
+        return null ?? auth()->user();
+    }
+}
