+++ composer.json
@@ -27,7 +27,10 @@
             "App\\": "app/",
             "Database\\Factories\\": "database/factories/",
             "Database\\Seeders\\": "database/seeders/"
-        }
+        },
+        "files": [
+            "app/Support/functions.php"
+        ]
     },
     "autoload-dev": {
         "psr-4": {
